requirements.txt
torch==2.0.1 
torchvision==0.15.2  
snntorch==0.5.1  
numpy==1.24.3
Pillow==9.4.0
matplotlib==3.7.1
